fix: defer load_app_config() to prevent circular import on submodule access
Importing any vibetuner.frontend submodule (e.g. vibetuner.frontend.oauth)
triggered __init__.py which eagerly called load_app_config(). This caused
circular imports when user route modules imported from vibetuner.frontend.*
during tune.py initialization.

Move the app factory into frontend/application.py and make __init__.py a
thin lazy re-export layer using __getattr__. Submodule imports no longer
trigger load_app_config().

Closes #1495
