+++ vibetuner-py/tests/unit/test_circular_import.py
@@ -1,6 +1,6 @@
 # ABOUTME: Tests that render_template is importable without triggering the frontend app factory.
 # ABOUTME: Regression tests for circular import issue #930.
-# ruff: noqa: S101, S603, S607
+# ruff: noqa: S101, S603
 import subprocess
 import sys
 import textwrap
@@ -73,6 +73,43 @@ class TestCircularImportPrevention:
 
         assert "OK" in result.stdout, (
             f"Importing from vibetuner triggered load_app_config():\n"
+            f"stderr: {result.stderr}"
+        )
+
+    def test_frontend_submodule_importable_without_load_app_config(self):
+        """Importing vibetuner.frontend.oauth must not trigger load_app_config().
+
+        When user route modules import from vibetuner.frontend.oauth (e.g., for
+        WebUser or register_oauth_provider), it must not trigger the app factory.
+        Otherwise a circular import occurs: __init__ -> load_app_config -> tune.py
+        -> user routes -> vibetuner.frontend.oauth -> __init__ (still initializing).
+
+        See: https://github.com/alltuner/vibetuner/issues/1495
+        """
+        code = textwrap.dedent("""\
+            import unittest.mock as mock
+            import vibetuner.loader
+
+            mock.patch.object(
+                vibetuner.loader,
+                "load_app_config",
+                side_effect=RuntimeError("load_app_config called during import"),
+            ).start()
+
+            from vibetuner.frontend.oauth import WebUser  # noqa: F401
+
+            print("OK")
+        """)
+
+        result = subprocess.run(
+            [sys.executable, "-c", code],
+            capture_output=True,
+            text=True,
+            timeout=30,
+        )
+
+        assert "OK" in result.stdout, (
+            f"Importing from vibetuner.frontend.oauth triggered load_app_config():\n"
             f"stderr: {result.stderr}"
         )
